ams: fix fatal error context

+++ libraries/libvapours/include/vapours/ams/ams_fatal_error_context.hpp
@@ -47,8 +47,8 @@ namespace ams::impl {
         u64 pc;
         u64 module_base;
         u32 pstate;
-        u32 afsr0;
-        u32 afsr1;
+        u64 afsr0;
+        u64 afsr1;
         u32 esr;
         u64 far;
         u64 report_identifier; /* Normally just system tick. */
